+++ sdk/resourcemanager/appcontainers/armappcontainers/client_factory.go
@@ -132,6 +132,11 @@ func (c *ClientFactory) NewManagedEnvironmentDiagnosticsClient() *ManagedEnviron
 	return subClient
 }
 
+func (c *ClientFactory) NewManagedEnvironmentUsagesClient() *ManagedEnvironmentUsagesClient {
+	subClient, _ := NewManagedEnvironmentUsagesClient(c.subscriptionID, c.credential, c.options)
+	return subClient
+}
+
 func (c *ClientFactory) NewManagedEnvironmentsClient() *ManagedEnvironmentsClient {
 	subClient, _ := NewManagedEnvironmentsClient(c.subscriptionID, c.credential, c.options)
 	return subClient
@@ -154,5 +159,10 @@ func (c *ClientFactory) NewNamespacesClient() *NamespacesClient {
 
 func (c *ClientFactory) NewOperationsClient() *OperationsClient {
 	subClient, _ := NewOperationsClient(c.credential, c.options)
+	return subClient
+}
+
+func (c *ClientFactory) NewUsagesClient() *UsagesClient {
+	subClient, _ := NewUsagesClient(c.subscriptionID, c.credential, c.options)
 	return subClient
 }