+++ sdk/resourcemanager/appcontainers/armappcontainers/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Release History
 
+## 2.1.0-beta.1 (2023-08-28)
+### Features Added
+
+- New function `*ClientFactory.NewManagedEnvironmentUsagesClient() *ManagedEnvironmentUsagesClient`
+- New function `*ClientFactory.NewUsagesClient() *UsagesClient`
+- New function `*ContainerAppsAPIClient.GetCustomDomainVerificationID(context.Context, *ContainerAppsAPIClientGetCustomDomainVerificationIDOptions) (ContainerAppsAPIClientGetCustomDomainVerificationIDResponse, error)`
+- New function `NewManagedEnvironmentUsagesClient(string, azcore.TokenCredential, *arm.ClientOptions) (*ManagedEnvironmentUsagesClient, error)`
+- New function `*ManagedEnvironmentUsagesClient.NewListPager(string, string, *ManagedEnvironmentUsagesClientListOptions) *runtime.Pager[ManagedEnvironmentUsagesClientListResponse]`
+- New function `NewUsagesClient(string, azcore.TokenCredential, *arm.ClientOptions) (*UsagesClient, error)`
+- New function `*UsagesClient.NewListPager(string, *UsagesClientListOptions) *runtime.Pager[UsagesClientListResponse]`
+- New struct `AppInsightsConfiguration`
+- New struct `BlobStorageTokenStore`
+- New struct `DataDogConfiguration`
+- New struct `DestinationsConfiguration`
+- New struct `EncryptionSettings`
+- New struct `IngressPortMapping`
+- New struct `ListUsagesResult`
+- New struct `LogsConfiguration`
+- New struct `MetricsConfiguration`
+- New struct `OpenTelemetryConfiguration`
+- New struct `TokenStore`
+- New struct `TracesConfiguration`
+- New struct `Usage`
+- New struct `UsageName`
+- New field `EncryptionSettings` in struct `AuthConfigProperties`
+- New field `AdditionalPortMappings` in struct `Ingress`
+- New field `TokenStore` in struct `Login`
+- New field `AppInsightsConfiguration`, `OpenTelemetryConfiguration` in struct `ManagedEnvironmentProperties`
+
+
 ## 2.0.0 (2023-08-25)
 ### Breaking Changes